Add retrasmission events, add event type flags and rename ebpftest tool
The new tool (i.e. toolebpflow) adds a command line interface to the
previous one.
