+++ ebpflow.ebpf
@@ -42,7 +42,7 @@ BPF_HASH(currsock, u32, struct sock_stats);
 // Retransmissions
 struct retr_entry {
   u64 ts;
-  u32 count;
+  u16 count;
 };
 #define BPF_LRU_HASH3(_name, _key_type, _leaf_type) BPF_TABLE("lru_hash", _key_type, _leaf_type, _name, 10240)
 BPF_LRU_HASH3(retr_table, u32, struct retr_entry);
@@ -52,10 +52,22 @@ BPF_LRU_HASH3(retr_table, u32, struct retr_entry);
 
 #define COMMAND_LEN  TASK_COMM_LEN
 
+/*
+ * Events types are forged as follows:
+ *  I_digit (=1): init events (e.g. connection creation)
+ *          (=2): update events on existing connection
+ *          (=3): connection closing
+ *  II_digit (=0): tcp events
+ *           (=1): udp events
+ *  III_digit: discriminate the single event
+ */
 typedef enum {
-  INIT = 0, // e.g TCP accept
-  END = 1, // e.g. TCP close
-  UPDATE = 2, // e.g. UDP send
+  eTCP_ACPT = 100,
+  eTCP_CONN = 101,
+  eUDP_RECV = 210,
+  eUDP_SEND = 211,
+  eTCP_RETR = 200,
+  eTCP_CLOSE = 300,
 } event_type;
 
 struct taskInfo {
@@ -71,6 +83,7 @@ struct netInfo {
   u16 dport;
   u8  proto;
   u32 latency_usec;
+  u16 retransmissions;
 };
 
 // separate data structs for ipv4 and ipv6
@@ -91,7 +104,7 @@ typedef struct {
   char ifname[IFNAMSIZ];
   struct timeval event_time;
   u_int8_t ip_version:4, sent_packet:4;
-  u_int8_t etype, return_code;
+  u16 etype;
 
   union {
     struct ipv4_data_t v4;
@@ -351,13 +364,13 @@ static int trace_connect_return(struct pt_regs *ctx) {
   int ret = PT_REGS_RC(ctx); // return value
   struct sock_stats *s;
   u32 tid = (bpf_get_current_pid_tgid() >> 32) & 0xFFFFFFFF;
-  eBPFevent event = { .ip_version = 4 };
+  eBPFevent event = { .etype = eTCP_CONN, .ip_version = 4 };
 
   s = currsock.lookup(&tid);
   if(s == NULL)
     return(0); // missed entry
 
-  if(ret == 0) {
+  if(ret == 0) { /* Capture failures */
     fill_event(ctx, &event, s->sk, NULL, s->ts, IPPROTO_TCP, 0 /* don't swap */);
     ebpf_events.perf_submit(ctx, &event, sizeof(eBPFevent));
   }
@@ -385,7 +398,7 @@ int trace_tcp_accept(struct pt_regs *ctx) {
   struct sock *newsk = (struct sock *)PT_REGS_RC(ctx);
 
   if(newsk != NULL) {
-    eBPFevent event = { .ip_version = 4 };
+    eBPFevent event = { .etype = eTCP_ACPT, .ip_version = 4 };
 
     fill_event(ctx, &event, newsk, NULL, 0, IPPROTO_TCP, 1 /* swap */);
     ebpf_events.perf_submit(ctx, &event, sizeof(eBPFevent));
@@ -400,7 +413,7 @@ int trace_tcp_accept(struct pt_regs *ctx) {
 int trace_tcp_close(struct pt_regs *ctx, struct sock *sk, int state) {
   if (state!=BPF_TCP_CLOSE) return 0;
   
-  eBPFevent event = { .etype = END, .ip_version = 6 };
+  eBPFevent event = { .etype = eTCP_CLOSE };
   
   fill_event(ctx, &event, sk, NULL, 0, IPPROTO_TCP, 0);
   ebpf_events.perf_submit(ctx, &event, sizeof(eBPFevent));
@@ -409,39 +422,45 @@ int trace_tcp_close(struct pt_regs *ctx, struct sock *sk, int state) {
 
 /* ******************************************* */
 
-// int trace_tcp_retransmit_skb (struct pt_regs *ctx, struct sock *sk) {
-//   struct retr_entry newentry = {};
-//   struct KernelData event_data = {};
-//   struct retr_entry *e;
-//   u32 tid;
-//   u64 diff, ts;
+int trace_tcp_retransmit_skb (struct pt_regs *ctx, struct sock *sk) {
+  struct retr_entry newentry = {};
+  eBPFevent event = { .etype = eTCP_RETR };
+  struct retr_entry *e;
+  u32 tid;  
+  u64 diff, ts;
   
-//   tid = (bpf_get_current_pid_tgid() >> 32) & 0xFFFFFFFF;  
-//   ts = bpf_ktime_get_ns();
+  tid = (bpf_get_current_pid_tgid() >> 32) & 0xFFFFFFFF;  
+  ts = bpf_ktime_get_ns();
 
-//   if ((e = retr_table.lookup(&tid))==NULL) {
-//     newentry.ts = ts;
-//     newentry.count = 1;
-//     retr_table.update(&tid, &newentry);
+  if ((e = retr_table.lookup(&tid))==NULL) {
+    newentry.ts = ts;
+    newentry.count = 1;
+    retr_table.update(&tid, &newentry);
 
-//     return -1;
-//   } else {
-//     e->count += 1;
-//     diff = ts - (e->ts);
-//     if (diff > 1000000000 /* 1 sec */) {
-//       event_data.netstat.retransmissions = e->count;
-//       fill_event(ctx, sk, NULL, &event_data, TCP_RETR);
-//       output_buffer.perf_submit(ctx, &event_data, sizeof(struct KernelData));
+    return -1;
+  } else {
+    e->count += 1;
+    diff = ts - (e->ts);
+    if (diff > 1000000000 /* 1 sec */) {
+      fill_event(ctx, &event, sk, NULL, 0, IPPROTO_TCP, 0);
+      // Setting retransmission count before sendig to user space
+      if (event.ip_version == 4){
+        event.event.v4.net.retransmissions = e->count;
+      }
+      else {
+        event.event.v6.net.retransmissions = e->count;
+      }
+      ebpf_events.perf_submit(ctx, &event, sizeof(eBPFevent));
 
-//       // Resetting timer
-//       newentry.ts = ts;
-//       newentry.count = e->count;
-//       retr_table.update(&tid, &newentry);
-//     }
-//   }
+      // Resetting timer
+      newentry.ts = ts;
+      newentry.count = e->count;
+      retr_table.update(&tid, &newentry);
+    }
+  }
 
-//   return 0;
-// }
+  return 0;
+}
 
 
 /* *********************** UDP *************************** */
@@ -516,7 +535,7 @@ static inline int udp_packet_trace(void *ctx, struct sk_buff* skb, u_int8_t sent
   // Compute MAC header address
   char* head;
   u16 mac_header;
-  eBPFevent event = { .sent_packet = sent_packet };
+  eBPFevent event = { .etype = eUDP_SEND, .sent_packet = sent_packet };
   u8 offset, l4proto, ip_version;
   char* ip_header_address;
   struct udphdr *udphdr;
@@ -672,7 +691,7 @@ int trace_netif_tx_entry(struct netif_rx_read_args *args) {
 }
 
 int trace_receive_v4(struct pt_regs *ctx, struct sock *sk) {
-  eBPFevent event = { .ip_version = 4 };
+  eBPFevent event = { .etype = eUDP_RECV, .ip_version = 4 };
 
   if(fill_event(ctx, &event, sk, NULL, bpf_ktime_get_ns(), IPPROTO_UDP, 0 /* don't swap */) == 0)
     ebpf_events.perf_submit(ctx, &event, sizeof(eBPFevent));
