+++ toolebpflow.cpp
@@ -23,23 +23,16 @@
 
 using namespace std;
 
-/* TODO: add vlan, remove newlines, defgault docker, dividi in .h .c, pulisci con variabili in cima */
-
 int gDOCKER_ENABLE=1;
 int gRUNNING = 1;
 
 void help();
-char* intoaV4(unsigned int addr, char* buf, u_short bufLen);
+static char* intoaV4(unsigned int addr, char* buf, u_short bufLen);
 char* intoaV6(unsigned __int128 addr, char* buf, u_short bufLen);
-void etype2str (int t_proto, char* t_buffer, int t_size);
-static void HandleEvent(void* t_bpfctx, void* t_data, int t_datasize);
-string LoadEBPF(string t_filepath);
-int AttachWrapper(ebpf::BPF* bpf, string t_kernel_fun, string t_ebpf_fun, bpf_probe_attach_type attach_type);
-static int attachEBPFTracepoint(ebpf::BPF *bpf, const char *tracepoint, const char *probe_func);
-int parse_uargs(int argc, char const *argv[]);
-void print_usage();
-static void HandleTermination(int t_s=0);
-
+void event_summary (eBPFevent* e, char* t_buffer, int t_size);
+static void handleTermination(int t_s=0);
+static void verboseHandleEvent(void* t_bpfctx, void* t_data, int t_datasize);
+static void ebpfHandler(void* t_bpfctx, void* t_data, int t_datasize);
 
 /* **************************** */
 // ===== ===== MAIN ===== ===== //
@@ -52,21 +45,23 @@ static const struct option long_opts[] = {
 	{ "in", 0, 0, 'i' },
 	{ "out", 0, 0, 'o' },
 	{ "docker", 0, 0, 'd' },
-  { "help", 0, 0, 'h' }
+  { "help", 0, 0, 'h' },
+  { "verbose", 0, 0, 'v' }
 };
 
 int main(int argc, char **argv) { 
   ebpfRetCode rc;
   void *ebpf;
+  void (*handler)(void*, void*, int) = ebpfHandler;
 
-  signal(SIGINT, HandleTermination);
+  signal(SIGINT, handleTermination);
 
   // Argument Parsing ----- //
   int ch;
   short flags = 0;
   gDOCKER_ENABLE=1;
   while ((ch = getopt_long(argc, argv,
-				                  "rcutiodh",
+				                  "rcutiodvh",
                           long_opts, NULL)) != EOF) {
       switch (ch) {
         case 'u':
@@ -86,6 +81,9 @@ int main(int argc, char **argv) {
           break;
         case 'd': 
           gDOCKER_ENABLE=1;        
+          break;
+        case 'v':
+          handler = verboseHandleEvent;
           break;
         default:
           help();
@@ -120,7 +118,7 @@ int main(int argc, char **argv) {
    "Legacy API"
 #endif
    );
-  ebpf = init_ebpf_flow(NULL, HandleEvent, &rc, flags);
+  ebpf = init_ebpf_flow(NULL, handler, &rc, flags);
 
   if(!ebpf) {
     printf("Unable to initialize libebpfflow: %s\n", ebpf_print_error(rc));
@@ -142,13 +140,14 @@ void help() {
   printf(
     "Usage: ebpflow [ OPTIONS ] \n"
     "   -h, --help      display this message \n"
-    "   -t, --tcp       trace TCP only events \n"
-    "   -u, --udp       trace UDP only events \n"
-    "   -i, --in        trace only incoming events (i.e. TCP accept and UDP receive) \n"
-    "   -i, --in        trace only outgoing events (i.e. TCP connect and UDP send) \n"
-    "   -r, --retr      trace only retransmissions events \n"
-    "   -c, --tcpclose  trace only tcp close events \n"
+    "   -t, --tcp       TCP events \n"
+    "   -u, --udp       UDP events \n"
+    "   -i, --in        incoming events (i.e. TCP accept and UDP receive) \n"
+    "   -i, --in        outgoing events (i.e. TCP connect and UDP send) \n"
+    "   -r, --retr      retransmissions events \n"
+    "   -c, --tcpclose  TCP close events \n"
     "   -d, --docker    gather additional information concerning containers \n"
+    "   -v, --verbose   vebose formatting"
     "(default: every event is shown) \n"
     "Note: please run as root \n"
   );
@@ -158,7 +157,7 @@ void help() {
 /* ******************************************** */
 // ===== ===== IP ADDRESS TO STRING ===== ===== //
 /* ******************************************** */
-char* intoaV4(unsigned int addr, char* buf, u_short bufLen) {
+static char* intoaV4(unsigned int addr, char* buf, u_short bufLen) {
   char *cp, *retStr;
   int n;
 
@@ -167,24 +166,17 @@ char* intoaV4(unsigned int addr, char* buf, u_short bufLen) {
 
   n = 4;
   do {
-    // Taking last byte
     u_int byte = addr & 0xff;
 
-    // Printing first cipher
     *--cp = byte % 10 + '0';
     byte /= 10;
-    // Checking if there are more ciphers
     if(byte > 0) {
-      // Writing second cipher
       *--cp = byte % 10 + '0';
       byte /= 10;
-      // Writing third cipher
       if(byte > 0)
-        *--cp = byte + '0';
+  *--cp = byte + '0';
     }
-    // Adding '.' character between decimals
     *--cp = '.';
-    // Shifting address of one byte (next step we'll take last byte)
     addr >>= 8;
   } while (--n > 0);
 
@@ -194,12 +186,11 @@ char* intoaV4(unsigned int addr, char* buf, u_short bufLen) {
   return(retStr);
 }
 
-char* intoaV6(unsigned __int128 addr, char* buf, u_short bufLen) {
-  char *ret = (char*)inet_ntop(AF_INET6, &addr, buf, bufLen);
+static char* intoaV6(void *addr, char* buf, u_short bufLen) {
+  char *ret = (char*)inet_ntop(AF_INET6, addr, buf, bufLen);
 
-  if(ret == NULL) {
+  if(ret == NULL)
     buf[0] = '\0';
-  }
 
   return(buf);
 }
@@ -209,38 +200,29 @@ char* intoaV6(unsigned __int128 addr, char* buf, u_short bufLen) {
 // ===== ===== CALLBACK HANDLERS ===== ===== //
 /* ***************************************** */
 void event_summary (eBPFevent* e, char* t_buffer, int t_size) {
-  __u8 net_proto;
-  __u8 sent_packet = e->sent_packet;
-  __u8 etype = e->etype;
-
-  if (e->ip_version == 4) {
-    net_proto = e->event.v4.net.proto;
-  }
-  else {
-    net_proto = e->event.v4.net.proto;
-  }
-
-  if (etype == END) {
-    strncpy(t_buffer, "TCP/close",  t_size);
-  }
-  else if (net_proto == IPPROTO_TCP) {
-    if (sent_packet) 
-      strncpy(t_buffer, "TCP/conn",  t_size);
-    else 
+  switch (e->etype) {
+    case eTCP_ACPT:
       strncpy(t_buffer, "TCP/acpt",  t_size);
-  }
-  else if (net_proto == IPPROTO_UDP) {
-    if (sent_packet) 
-      strncpy(t_buffer, "UDP/snd",  t_size);
-    else 
-      strncpy(t_buffer, "UDP/rcv",  t_size);
-  } 
-  else {
-    strncpy(t_buffer, "?",  t_size);
+      break;
+    case eTCP_CONN:
+      strncpy(t_buffer, "TCP/conn",  t_size);
+      break;
+    case eTCP_CLOSE:
+      strncpy(t_buffer, "TCP/close",  t_size);
+      break;
+    case eTCP_RETR:
+      strncpy(t_buffer, "TCP/retr",  t_size);
+      break;
+    case eUDP_SEND:
+      strncpy(t_buffer, "UDP/send",  t_size);
+      break;
+    case eUDP_RECV:
+      strncpy(t_buffer, "UDP/recv",  t_size);
+      break;
   }
 }
 
-static void HandleEvent(void* t_bpfctx, void* t_data, int t_datasize) {
+static void verboseHandleEvent(void* t_bpfctx, void* t_data, int t_datasize) {
   char event_type_str[17];
   eBPFevent *e = (eBPFevent*)t_data;
 
@@ -295,8 +277,8 @@ static void HandleEvent(void* t_bpfctx, void* t_data, int t_datasize) {
       event.ifname,
       event.ip_version,
       event_type_str,
-      intoaV6(htonl(ipv6_event->saddr), buf1, sizeof(buf1)), ipv6_event->net.sport,
-      intoaV6(htonl(ipv6_event->daddr), buf2, sizeof(buf2)), ipv6_event->net.dport);
+      intoaV6(&ipv6_event->saddr, buf1, sizeof(buf1)), ipv6_event->net.sport,
+      intoaV6(&ipv6_event->saddr, buf2, sizeof(buf2)), ipv6_event->net.dport);
     // IPv6 Latency if available
     if(strcmp(event_type_str, "TCP/conn") == 0)
       printf("\t [latency: %.2f msec] (netstat) \n", ((float)ipv6_event->net.latency_usec)/(float)1000);
@@ -309,7 +291,90 @@ static void HandleEvent(void* t_bpfctx, void* t_data, int t_datasize) {
   ebpf_free_event(&event);
 }
 
-static void HandleTermination(int t_s) {
+/* ***************************************************** */
+/* ***************************************************** */
+
+static void IPV4Handler(void* t_bpfctx, eBPFevent *e, struct ipv4_kernel_data *event) {
+  char buf1[32], buf2[32];
+
+  printf("[%s][%s][IPv4/%s][pid/tid: %u/%u [%s], uid/gid: %u/%u][father pid/tid: %u/%u [%s], uid/gid: %u/%u][addr: %s:%u <-> %s:%u]",
+   e->ifname, e->sent_packet ? "Sent" : "Rcvd",
+   (event->net.proto == IPPROTO_TCP) ? "TCP" : "UDP",
+   e->proc.pid, e->proc.tid,
+   (e->proc.full_task_path == NULL) ? e->proc.task : e->proc.full_task_path,
+   e->proc.uid, e->proc.gid,
+   e->father.pid, e->father.tid,
+   (e->father.full_task_path == NULL) ? e->father.task : e->father.full_task_path,
+   e->father.uid, e->father.gid,
+   intoaV4(htonl(event->saddr), buf1, sizeof(buf1)), event->net.sport,
+   intoaV4(htonl(event->daddr), buf2, sizeof(buf2)), event->net.dport);
+
+  if (e->docker != NULL) printf("[docker: %s/%s]", e->cgroup_id, e->docker->dname);
+  if (e->kube !=  NULL) printf("[kube pod/ns: %s/%s]", e->kube->pod, e->kube->ns);
+
+  if(event->net.proto == IPPROTO_TCP)
+    printf("[latency: %.2f msec]", ((float)event->net.latency_usec)/(float)1000);
+}
+
+
+static void IPV6Handler(void* t_bpfctx, eBPFevent *e, struct ipv6_kernel_data *event) {
+  char buf1[32], buf2[32];
+  unsigned long long high = (event->saddr >> 64) &0xFFFFFFFFFFFFFFFF;
+  unsigned long long low = event->saddr & 0xFFFFFFFFFFFFFFFF;
+
+  printf("[%s][%s][IPv6/%s][pid/tid: %u/%u (%s [%s]), uid/gid: %u/%u][father pid/tid: %u/%u (%s [%s]), uid/gid: %u/%u][addr: %s:%u <-> %s:%u]",
+   e->ifname, e->sent_packet ? "S" : "R",
+   (event->net.proto == IPPROTO_TCP) ? "TCP" : "UDP", e->proc.pid, e->proc.tid,
+   e->proc.task, (e->proc.full_task_path == NULL) ? e->proc.task : e->proc.full_task_path,
+   e->proc.uid, e->proc.gid,
+   e->father.pid, e->father.tid,
+   e->proc.task, (e->father.full_task_path == NULL) ? e->father.task : e->father.full_task_path,
+   e->father.uid, e->father.gid,
+   intoaV6(&event->saddr, buf1, sizeof(buf1)),
+   event->net.sport,
+   intoaV6(&event->daddr, buf2, sizeof(buf2)),
+   event->net.dport);
+
+  if (e->docker != NULL) printf("[docker: %s/%s]", e->cgroup_id, e->docker->dname);
+  if (e->kube != NULL) printf("[kube pod/ns: %s/%s]", e->kube->pod, e->kube->ns);
+
+  if(event->net.proto == IPPROTO_TCP)
+    printf("[latency: %.2f msec]", ((float)event->net.latency_usec)/(float)1000);
+}
+
+
+static void ebpfHandler(void* t_bpfctx, void* t_data, int t_datasize) {
+  eBPFevent *e = (eBPFevent*)t_data;
+  eBPFevent event;
+  struct timespec tp;
+
+  memcpy(&event, e, sizeof(eBPFevent)); /* Copy needed as ebpf_preprocess_event will modify the memory */
+
+  ebpf_preprocess_event(&event, 1);
+
+  clock_gettime(CLOCK_MONOTONIC, &tp);
+
+#if 0
+  printf("[latency %.1f usec] ",
+   (float)(tp.tv_nsec-(event.ktime % 1000000000))/(float)1000);
+#endif
+
+  printf("%u.%06u ",
+   (unsigned int)event.event_time.tv_sec,
+   (unsigned int)event.event_time.tv_usec);
+  
+  if(event.ip_version == 4)
+    IPV4Handler(t_bpfctx, &event, &event.event.v4);
+  else
+    IPV6Handler(t_bpfctx, &event, &event.event.v6);
+
+  printf("\n");
+  ebpf_free_event(&event);
+}
+
+
+
+static void handleTermination(int t_s) {
   printf("\r* Terminating * \n");
   gRUNNING = 0;
 }
